requirements.txt
# Core GUI framework
PyQt5>=5.15.0

# Data handling and scientific computing
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
PIL>=8.3.0
Pillow>=8.3.0

# Hardware integration
pylsl>=1.16.0
paramiko>=2.11.0
pyxdf>=1.16.0
pupil-labs-realtime-api>=1.0.0

# Windows-specific automation (optional)
pywinauto>=0.6.8; sys_platform == "win32"

# Other dependencies
PyYAML>=6.0